camera/test_camera.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for mozCameras</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="doTest()">

<script class="testbody" type="text/javascript;version=1.7">

var nav = SpecialPowers.wrap(navigator);
var currentDate;

function doTest() {
try {
  navigator.mozCameras;
  ok(false, "navigator.mozCameras should throw in regular mochitest");
} catch(e) {
  ok(true, "navigator.mozCameras should throw in regular mochitest");
}

try {
  nav.mozCameras.previewSizes;
  ok(false, "chromeWindow.navigator.mozCameras.previewSizes should throw");
} catch(e) {
  ok(true, "chromeWindow.navigator.mozCameras.previewSizes should throw");
}
SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

</script>
</body>

</html>
